Implement batch clan file processing
You can point the script to a directory full of clan files and it
will process all of them. The output csv are automatically named based
on the clan file that they correspond to. You need to provide an output
directory rather than a single csv filepath now. All the csv's will be
output to this directory.