Run clanstats on all 6/7 month files, then run batch_clanstats on them
Running all the 6 and 7 month files through clanstats brought out a
number of bugs in the script, including many missing speaker codes that
had to be added to the dictionaries. It also brought out a number of
mistakes in the clan files that had to be fixed before going further
(things like MOt and fat, but also codes like TOYS and ADU2 which are 4
letters long (in 08_07 and 09_06 respectively). These mistakes have
been fixed in the local copies of the files in this repository, but
not in the originals on Opus. I kept a record of where each one was so
we can go back and deal with them in the originals).

The aggregate counts are output to a file called aggregate_stats.txt.
This file has all the information necessary to fill a confusion matrix
for the lena speaker classifiers.

Next commits will fix bugs that are preventing analysis of certain
files. Also going to merge the .cha parsing code from parse_clan2
so the script can work on the new format as well.